requirements.txt
langgraph
langchain-community==0.3.18
langchain-text-splitters
langchain_core
langchain-groq==0.3.2
python-dotenv
pydantic
python-multipart
streamlit
ipykernel
langchain-openai
langchain-google-genai
fastapi
uvicorn
langchain-chroma
chromadb
twilio
SpeechRecognition
PyAudio
streamlit-webrtc
streamlit_audio_recorder
soundfile